Update tuple to list

+++ pygmt/src/surface.py
@@ -98,7 +98,7 @@ def surface(data=None, x=None, y=None, z=None, **kwargs):
         This is the final convergence limit at the desired grid spacing;
         for intermediate (coarser) grids the effective convergence limit is
         divided by the grid spacing multiplier.
-    maxradius : int or str
+    maxradius : float or str
         Optional. After solving for the surface, apply a mask so that nodes
         farther than ``maxradius`` away from a data constraint are set to NaN
         [Default is no masking]. Append a distance unit (see
